tests/wpt/tests/websockets/interfaces/WebSocket/extensions/001.html

<!doctype html>
<title>WebSockets: getting extensions in connecting</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function(t) {
  // The extensions attribute must initially return the empty string
  assert_equals((new WebSocket(SCHEME_DOMAIN_PORT+'/empty-message')).extensions, '');
});
</script>
